backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
torch==2.2.0
torchvision==0.17.0
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.0
Pillow==10.2.0
opencv-python==4.9.0.80
google-generativeai==0.4.0
joblib==1.3.2
python-dotenv==1.2.2
